Supplier Module Done

+++ Modules/Supplier/Http/Controllers/SupplierController.php
@@ -3,7 +3,6 @@
 namespace Modules\Supplier\Http\Controllers;
 
 use Exception;
-use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Modules\Supplier\Entities\Supplier;
@@ -11,7 +10,7 @@ use App\Http\Controllers\BaseController;
 use Modules\Account\Entities\Transaction;
 use Modules\Account\Entities\ChartOfAccount;
 use Modules\Supplier\Http\Requests\SupplierFormRequest;
-
+use Illuminate\Support\Facades\App;
 
 class SupplierController extends BaseController
 {
@@ -46,9 +45,7 @@ class SupplierController extends BaseController
                 if (!empty($request->email)) {
                     $this->model->setEmail($request->email);
                 }
-                if (!empty($request->type)) {
-                    $this->model->setType($request->type);
-                }
+
                 if (!empty($request->status)) {
                     $this->model->setStatus($request->status);
                 }
@@ -73,17 +70,13 @@ class SupplierController extends BaseController
                         $row[] = row_checkbox($value->id);
                     }
                     $balance = $this->model->supplier_balance($value->id);
-                    $balance = ($currency_position == 1) ? $currency_symbol.' '.$balance : $balance.' '.$currency_symbol;
-                    $row[] = $no;
-                    $row[] = $value->company_name ? $value->name.' ('.$value->company_name.')' : $value->name;
+                    $row[] = translate($no,App::getLocale());
+                    $row[] = $value->name.'<br><b>Mobile No.: </b>'.$value->mobile.($value->email ? '<br><b>Email: </b>'.$value->email.')' : '').($value->company_name ? '<br><b>Company Name: </b>'.$value->company_name : '');
                     $row[] = $value->address;
-                    $row[] = $value->mobile;
-                    $row[] = $value->email;
                     $row[] = $value->city;
                     $row[] = $value->zipcode;
-                    $row[] = SUPPLIER_TYPE_LABEL[$value->type];
                     $row[] = permission('supplier-edit') ? change_status($value->id,$value->status, $value->name) : STATUS_LABEL[$value->status];
-                    $row[] = $balance;
+                    $row[] = translate(number_format($balance,2,'.',','),App::getLocale());
                     $row[] = action_button($action);//custom helper function for action button
                     $data[] = $row;
                 }
@@ -107,38 +100,29 @@ class SupplierController extends BaseController
                     $collection = $this->track_data($collection,$request->update_id);
                     $supplier   = $this->model->updateOrCreate(['id'=>$request->update_id],$collection->all());
                     $output     = $this->store_message($supplier, $request->update_id);
-                    if(empty($request->update_id))
+                    if($request->update_id)
                     {
+                        $supplier_coa = ChartOfAccount::where(['supplier_id'=>$request->update_id])->first();
+                        if($supplier_coa)
+                        {
+                            $supplier_coa->update(['name'=>$request->name]);
+                        }else{
+                            $coa_max_code      = ChartOfAccount::where('level',3)->where('code','like','50201%')->max('code');
+                            $code              = $coa_max_code ? ($coa_max_code + 1) : '502010001';
+                            ChartOfAccount::create($this->coa_data($code,$request->name,$request->update_id));
+                        }
+                    }else{
                         $coa_max_code      = ChartOfAccount::where('level',3)->where('code','like','50201%')->max('code');
-                        $code              = $coa_max_code ? ($coa_max_code + 1) : $this->coa_head_code('default_supplier');
-                        $head_name         = $supplier->id.'-'.$supplier->name;
-                        $supplier_coa_data = $this->supplier_coa($code,$head_name,$supplier->id);
-
-                        $supplier_coa      = ChartOfAccount::create($supplier_coa_data);
+                        $code              = $coa_max_code ? ($coa_max_code + 1) : '50201000001';
+                        $supplier_coa      = ChartOfAccount::create($this->coa_data($code,$supplier->name,$supplier->id));
                         if(!empty($request->previous_balance))
                         {
                             if($supplier_coa){
-                                $this->previous_balance_add($request->previous_balance,$supplier_coa->id,$supplier->name);
-                            }
-                        }
-                    }else{
-                        $old_head_name = $request->update_id.'-'.$request->old_name;
-                        $new_head_name = $request->update_id.'-'.$request->name;
-                        $supplier_coa = ChartOfAccount::where(['name'=>$old_head_name,'supplier_id'=>$request->update_id])->first();
-                        if($supplier_coa)
-                        {
-                            $supplier_coa_id = $supplier_coa->id;
-                            $supplier_coa->update(['name'=>$new_head_name]);
-                            if(!empty($request->previous_balance) && !empty($request->old_previous_balance))
-                            {
-                                if($request->previous_balance != $request->old_previous_balance){
-                                    $this->previous_balance_update($request->previous_balance,$supplier_coa_id,$request->name);
-                                }
-                            }else{
-                                $this->previous_balance_add($request->previous_balance,$supplier_coa_id,$request->name);
+                                Transaction::insert($this->previous_balance_data($request->previous_balance,$supplier_coa->id,$supplier->name));
                             }
                         }
                     }
+                    $this->model->flushCache();
                     DB::commit();
                 } catch (Exception $e) {
                     DB::rollBack();
@@ -154,7 +138,7 @@ class SupplierController extends BaseController
     }
 
 
-    private function supplier_coa(string $code,string $head_name,int $supplier_id)
+    private function coa_data(string $code,string $head_name,int $supplier_id)
     {
         return [
             'code'              => $code,
@@ -164,7 +148,6 @@ class SupplierController extends BaseController
             'type'              => 'L',
             'transaction'       => 1,
             'general_ledger'    => 2,
-            'customer_id'       => null,
             'supplier_id'       => $supplier_id,
             'budget'            => 2,
             'depreciation'      => 2,
@@ -174,18 +157,18 @@ class SupplierController extends BaseController
         ];
     }
 
-    private function previous_balance_add($balance, int $supplier_coa_id, string $supplier_name) {
+    private function previous_balance_data($balance, int $supplier_coa_id, string $supplier_name) {
         if(!empty($balance) && !empty($supplier_coa_id) && !empty($supplier_name)){
-            $transaction_id = generator(10);
+            $transaction_id = date('ymdHi').rand(1,99);
             // supplier debit for previous balance
             $cosdr = array(
                 'chart_of_account_id' => $supplier_coa_id,
                 'voucher_no'          => $transaction_id,
                 'voucher_type'        => 'PR Balance',
                 'voucher_date'        => date("Y-m-d"),
-                'description'         => 'Supplier debit For '.$supplier_name,
-                'debit'               => $balance,
-                'credit'              => 0,
+                'description'         => 'Credit '.$balance.' for old purchase from supplier '.$supplier_name,
+                'debit'               => 0,
+                'credit'              => $balance,
                 'posted'              => 1,
                 'approve'             => 1,
                 'created_by'          => auth()->user()->name,
@@ -196,49 +179,18 @@ class SupplierController extends BaseController
                 'voucher_no'          => $transaction_id,
                 'voucher_type'        => 'PR Balance',
                 'voucher_date'        => date("Y-m-d"),
-                'description'         => 'Inventory credit For Old sale For '.$supplier_name,
-                'debit'               => 0,
-                'credit'              => $balance,
+                'description'         => 'Inventory dredit '.$balance.' for old purchase from supplier '.$supplier_name,
+                'debit'               => $balance,
+                'credit'              => 0,
                 'posted'              => 1,
                 'approve'             => 1,
                 'created_by'          => auth()->user()->name,
                 'created_at'          => date('Y-m-d H:i:s')
             );
 
-            Transaction::insert([
+            return [
                 $cosdr,$inventory
-            ]);
-        }
-    }
-
-    private function previous_balance_update($balance, int $supplier_coa_id, string $supplier_name) {
-        if(!empty($balance) && !empty($supplier_coa_id) && !empty($supplier_name)){
-
-            $supplier_pr_balance_data = Transaction::where(['chart_of_account_id' => $supplier_coa_id, 'voucher_type'=> 'PR Balance',])->first();
-
-            $voucher_no = $supplier_pr_balance_data->voucher_no;
-
-            $updated = $supplier_pr_balance_data->update([
-                'description'         => 'Supplier debit For '.$supplier_name,
-                'debit'               => $balance,
-                'modified_by'         => auth()->user()->name,
-                'updated_at'          => date('Y-m-d H:i:s')
-            ]);
-            if($updated)
-            {
-                Transaction::where([
-                    'chart_of_account_id' => DB::table('chart_of_accounts')->where('code', $this->coa_head_code('inventory'))->value('id'),
-                    'voucher_no'=> $voucher_no])
-                    ->update([
-                        'description'         => 'Inventory credit For Old sale For '.$supplier_name,
-                        'credit'              => $balance,
-                        'modified_by'         => auth()->user()->name,
-                        'updated_at'          => date('Y-m-d H:i:s')
-                    ]);
-                return true;
-            }else{
-                return false;
-            }
+            ];
         }
     }
 
@@ -246,7 +198,7 @@ class SupplierController extends BaseController
     {
         if($request->ajax()){
             if(permission('supplier-edit')){
-                $data   = $this->model->with('previous_balance')->findOrFail($request->id);
+                $data   = $this->model->findOrFail($request->id);
                 $output = $this->data_message($data); //if data found then it will return data otherwise return error message
             }else{
                 $output       = $this->unauthorized();
@@ -261,8 +213,26 @@ class SupplierController extends BaseController
     {
         if($request->ajax()){
             if(permission('supplier-delete')){
-                $result   = $this->model->find($request->id)->delete();
-                $output   = $this->delete_message($result);
+                DB::beginTransaction();
+                try {
+                    $total_purchase_data = DB::table('purchase_orders')->where('supplier_id',$request->id)->get()->count();
+                    if ($total_purchase_data > 0) {
+                        $output = ['status'=>'error','message'=> __('file.This data cannot delete because it is related with others data.')];
+                    } else {
+                        $supplier_coa_id = ChartOfAccount::where('supplier_id',$request->id)->first();
+                        if($supplier_coa_id){
+                            Transaction::where('chart_of_account_id',$supplier_coa_id->id)->delete();
+                            $supplier_coa_id->delete();
+                        }
+                        $result   = $this->model->find($request->id)->delete();
+                        $output   = $this->delete_message($result);
+                    }
+                    DB::commit();
+                    $this->model->flushCache();
+                } catch (Exception $e) {
+                   DB::rollBack();
+                   $output = ['status' => 'error','message' => $e->getMessage()];
+                }
             }else{
                 $output       = $this->unauthorized();
             }
@@ -276,8 +246,28 @@ class SupplierController extends BaseController
     {
         if($request->ajax()){
             if(permission('supplier-bulk-delete')){
-                $result   = $this->model->destroy($request->ids);
-                $output   = $this->bulk_delete_message($result);
+                DB::beginTransaction();
+                try {
+                    foreach ($request->ids as $id) {
+                        $total_purchase_data = DB::table('purchase_orders')->where('supplier_id',$id)->get()->count();
+                        if ($total_purchase_data > 0) {
+                            $output = ['status'=>'error','message'=> __('file.This data cannot delete because it is related with others data.')];
+                        } else {
+                            $supplier_coa_id = ChartOfAccount::where('supplier_id',$id)->first();
+                            if($supplier_coa_id){
+                                Transaction::where('chart_of_account_id',$supplier_coa_id->id)->delete();
+                                $supplier_coa_id->delete();
+                            }
+                        }
+                    }
+                    $result   = $this->model->destroy($request->ids);
+                    $output   = $this->delete_message($result);
+                    DB::commit();
+                    $this->model->flushCache();
+                } catch (Exception $e) {
+                   DB::rollBack();
+                   $output = ['status' => 'error','message' => $e->getMessage()];
+                } 
             }else{
                 $output       = $this->unauthorized();
             }
